requirements.txt
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
langchain-community
langchain-openai
streamlit
beautifulsoup4
lxml
